binding the post content to the post details page
