+++ data/src/main/java/com/ouchadam/loldr/data/deserialize/PostDeserializer.java
@@ -1,6 +1,5 @@
 package com.ouchadam.loldr.data.deserialize;
 
-import com.google.gson.JsonArray;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonElement;
@@ -9,82 +8,54 @@ import com.google.gson.JsonParseException;
 import com.ouchadam.loldr.data.Data;
 
 import java.lang.reflect.Type;
-import java.util.ArrayList;
-import java.util.List;
 
-class PostDeserializer implements JsonDeserializer<Data.Comments> {
+class PostDeserializer implements JsonDeserializer<Data.Post> {
 
     @Override
-    public Data.Comments deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
-        JsonObject commentsRootJson = json.getAsJsonArray().get(1).getAsJsonObject();
-        JsonArray rootCommentThread = commentsRootJson.get("data").getAsJsonObject().get("children").getAsJsonArray();
-
-        List<Data.Comment> comments = new ArrayList<>();
-        for (JsonElement jsonElement : rootCommentThread) {
-            JsonObject asJsonObject = jsonElement.getAsJsonObject();
-            comments.addAll(recurseOverAllComments(asJsonObject, new ArrayList<Data.Comment>(), 0));
-        }
-
-        return new Comments(comments);
+    public Data.Post deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+        JsonObject postJson = json.getAsJsonObject();
+        return new Post(
+                postJson.get("id").getAsString(),
+                postJson.get("title").getAsString(),
+                postJson.get("author").getAsString(),
+                postJson.get("subreddit").getAsString(),
+                postJson.get("score").getAsInt(),
+                postJson.get("num_comments").getAsInt(),
+                postJson.get("created_utc").getAsLong(),
+                postJson.get("saved").getAsBoolean(),
+                postJson.get("thumbnail").getAsString(),
+                postJson.get("url").getAsString(),
+                postJson.get("selftext").getAsString()
+        );
     }
 
-    private List<Data.Comment> recurseOverAllComments(JsonObject commentThread, List<Data.Comment> comments, int depth) {
-        String kind = commentThread.get("kind").getAsString();
-        if (kind.equals("more")) {
-            // TODO these type of comments require a different api call /api/morechildren taking the child ids
-            Data.Comment moreComment = moreComment(commentThread.get("data").getAsJsonObject(), depth);
-            comments.add(moreComment);
-            return comments;
-        }
-
-        JsonObject commentJson = commentThread.get("data").getAsJsonObject();
-        comments.add(parseComment(commentJson, depth));
-
-        JsonElement repliesRoot = commentJson.get("replies");
-        if (repliesRoot.isJsonPrimitive()) {
-            return comments;
-        }
-
-        JsonArray replies = repliesRoot.getAsJsonObject().get("data").getAsJsonObject().get("children").getAsJsonArray();
-
-        for (JsonElement reply : replies) {
-            comments.addAll(recurseOverAllComments(reply.getAsJsonObject(), new ArrayList<Data.Comment>(), depth + 1));
-        }
-
-        return comments;
-    }
-
-    private Data.Comment moreComment(JsonObject data, int depth) {
-        String id = data.get("id").getAsString();
-        String name = data.get("name").getAsString();
-        return new MoreComment(id, name, depth);
-    }
-
-    private Comment parseComment(JsonObject jsonComment, int depth) {
-        String commentId = jsonComment.get("id").getAsString().trim();
-        String commentBody = jsonComment.get("body").getAsString().trim();
-        String commentName = jsonComment.get("name").getAsString().trim();
-        long commentTimestamp = jsonComment.get("created_utc").getAsLong();
-        String commentAuthor = jsonComment.get("author").getAsString().trim();
-
-        return new Comment(commentId, commentBody, commentName, commentTimestamp, commentAuthor, depth);
-    }
-
-    private static class MoreComment implements Data.Comment {
+    private static class Post implements Data.Post {
 
         private final String id;
-        private final String name;
-        private final int depth;
+        private final String title;
+        private final String author;
+        private final String subreddit;
+        private final int score;
+        private final int commentCount;
+        private final long createdUtcTimeStamp;
+        private final boolean saved;
+        private final String thumbnail;
+        private final String url;
+        private final String content;
 
-        public MoreComment(String id, String name, int depth) {
+        public Post(String id, String title, String author, String subreddit, int score, int commentCount,
+                    long createdUtcTimeStamp, boolean saved, String thumbnail, String url, String content) {
             this.id = id;
-            this.name = name;
-            this.depth = depth;
-        }
-
-        @Override
-        public String getBody() {
-            throw new IllegalAccessError("This is a more comment! no body brah");
+            this.title = title;
+            this.author = author;
+            this.subreddit = subreddit;
+            this.score = score;
+            this.commentCount = commentCount;
+            this.createdUtcTimeStamp = createdUtcTimeStamp;
+            this.saved = saved;
+            this.thumbnail = thumbnail;
+            this.url = url;
+            this.content = content;
         }
 
         @Override
@@ -92,90 +63,56 @@ class PostDeserializer implements JsonDeserializer<Data.Comments> {
             return id;
         }
 
-        public String getName() {
-            return name;
+        @Override
+        public String getTitle() {
+            return title;
+        }
+
+        @Override
+        public String getSubreddit() {
+            return subreddit;
+        }
+
+        @Override
+        public int getScore() {
+            return score;
+        }
+
+        @Override
+        public int getCommentCount() {
+            return commentCount;
+        }
+
+        @Override
+        public long getCreatedDate() {
+            return createdUtcTimeStamp;
         }
 
         @Override
         public String getAuthor() {
-            throw new IllegalAccessError("This is a more comment! no author brah");
+            return author;
         }
 
         @Override
-        public int getDepth() {
-            return depth;
+        public boolean isSaved() {
+            return saved;
         }
 
         @Override
-        public boolean isMore() {
-            return true;
-        }
-    }
-
-    private static class Comments implements Data.Comments {
-
-        private final List<Data.Comment> comments;
-
-        public Comments(List<Data.Comment> comments) {
-            this.comments = comments;
+        public String getThumbnailUrl() {
+            return thumbnail;
         }
 
         @Override
-        public List<Data.Comment> getComments() {
-            return comments;
-        }
-    }
-
-    private static class Comment implements Data.Comment {
-
-        private final String id;
-        private final String body;
-        private final String name;
-        private final long createdUtc;
-        private final String commentAuthor;
-        private final int depth;
-
-        public Comment(String id, String body, String name, long createdUtc, String commentAuthor, int depth) {
-            this.id = id;
-            this.body = body;
-            this.name = name;
-            this.createdUtc = createdUtc;
-            this.commentAuthor = commentAuthor;
-            this.depth = depth;
+        public String getExternalLink() {
+            return url;
         }
 
         @Override
-        public String getBody() {
-            return body;
+        public String getContent() {
+            return content;
         }
 
-        @Override
-        public String getId() {
-            return id;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public long getCreatedUtc() {
-            return createdUtc;
-        }
-
-        @Override
-        public String getAuthor() {
-            return commentAuthor;
-        }
-
-        @Override
-        public int getDepth() {
-            return depth;
-        }
-
-        @Override
-        public boolean isMore() {
-            return false;
-        }
     }
 
 }
